test(comms): share the pending-queue and daily-digest environments
telegram-pending-queue-drain.test.ts (2,904 lines) and its sibling
restated queue-row literals per case, and daily-digest.test.ts and
daily-digest-generation.test.ts held a byte-identical digest test
environment that the clone detector measured at 484 duplicated lines.

Adds telegram-pending-queue.test-support.ts and
daily-digest.test-support.ts, converts the repeated queue literals to
factories plus case tables, and extends the existing
dispatch-telegram-alerts.test-support.ts rather than adding a third
overlapping helper. 264 cases pass with unchanged counts.

Net -207 lines.
