+++ worker/src/cron/__tests__/telegram-pending-queue-drain.test.ts
@@ -7,18 +7,11 @@ import {
 } from "../../lib/telegram-pending-provenance";
 import { createSqliteD1 } from "../../test-helpers/sqlite-d1";
 import { createLatestSchemaSqlite } from "../../test-helpers/latest-schema-sqlite";
-
-const DEFAULT_TELEGRAM_PENDING_D1_TABLES: MockTableConfig[] = [
-  { match: "WHERE delivery_state = 'sending'", rows: [] },
-  { match: "delivery_state = 'sent'", rows: [] },
-  { match: "processing_owner = ?", rows: [] },
-  { match: "SET attempts = attempts + 1", rows: [] },
-  { match: "AND delivery_state = 'sending'", rows: [] },
-  { match: "DELETE FROM telegram_preset_subscriptions", rows: [] },
-  { match: "WHERE chat_id = ?", rows: [] },
-  { match: "UPDATE telegram_recap_preferences", rows: [] },
-  { match: "UPDATE telegram_recap_targets", rows: [] },
-];
+import {
+  DEFAULT_TELEGRAM_PENDING_D1_TABLES,
+  insertPendingSqlite,
+  makeTelegramDeliveryResult,
+} from "./telegram-pending-queue.test-support";
 
 function mockD1(tables: MockTableConfig[] = []) {
   return createMockD1([...tables, ...DEFAULT_TELEGRAM_PENDING_D1_TABLES]);
@@ -102,58 +95,6 @@ function insertSubscriberSqlite(
       row.globalAlertLaunch ?? 0,
       row.globalAlertReserve ?? 0,
       row.globalAlertFreeze ?? 0,
-    );
-}
-
-function insertPendingSqlite(
-  sqlite: DatabaseSync,
-  row: {
-    id: number;
-    chatId: string;
-    html: string;
-    createdAt: number;
-    disableNotification?: number;
-    attempts?: number;
-    notBeforeAt?: number | null;
-    dedupeKey?: string | null;
-    chunkIndex?: number | null;
-    priority?: number | null;
-    sourceType?: string | null;
-    alertType?: string | null;
-    expiresAt?: number | null;
-    sourceEventId?: string | null;
-    alertScopeJson?: string | null;
-    preferenceGeneration?: number | null;
-    markupPolicyJson?: string | null;
-  },
-): void {
-  sqlite
-    .prepare(
-      `INSERT INTO telegram_pending_alerts (
-         id, chat_id, message_html, disable_notification, created_at, attempts,
-         not_before_at, dedupe_key, chunk_index, priority, source_type, alert_type, expires_at,
-         source_event_id, alert_scope_json, preference_generation, markup_policy_json
-       )
-       VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)`,
-    )
-    .run(
-      row.id,
-      row.chatId,
-      row.html,
-      row.disableNotification ?? 0,
-      row.createdAt,
-      row.attempts ?? 0,
-      row.notBeforeAt ?? null,
-      row.dedupeKey ?? null,
-      row.chunkIndex ?? 0,
-      row.priority ?? TELEGRAM_PENDING_PRIORITY.legacy,
-      row.sourceType ?? "legacy",
-      row.alertType ?? null,
-      row.expiresAt ?? null,
-      row.sourceEventId ?? null,
-      row.alertScopeJson ?? null,
-      row.preferenceGeneration ?? null,
-      row.markupPolicyJson ?? null,
     );
 }
